Remove explicit version dependency
Previously, the code depended on an explicit version of a Spring Boot
dependency.  As Spring Boot's version could change underneath it, this
fixed version was problematic.  In 0.5.0.M5, Spring Boot added the
ability to simply declare an artifact without a version, and the
suggested version (from Spring Boot) would be applied automatically.
This change removes the fixed version to take advantage of this
feature.